Resolves #15 by removing EPPlus Dependency
