+++ src/NetCore.Utilities.Spreadsheet/DependencyResolution/StartupExtensions.cs
@@ -14,7 +14,7 @@ namespace Microsoft.Extensions.DependencyInjection
         public static void UseIcgNetCoreUtilitiesSpreadsheet(this IServiceCollection services)
         {
             //Bind additional services
-            services.AddTransient<ISpreadsheetGenerator, EPPlusSpreadsheetGenerator>();
+            services.AddTransient<ISpreadsheetGenerator, OpenXmlSpreadsheetGenerator>();
             services.AddTransient<ISpreadsheetParser, OpenXmlSpreadsheetParser>();
         }
     }
